--- src/app/component/Button.js
@@ -1,11 +1,0 @@
-import font from "../font.module.css";
-
-export default function btnPrimary() {
-  return (
-    <div className=" bg-slate-300">
-      <p>
-        test
-      </p>
-    </div>
-  );
-}
