+++ src/app/globals.css
@@ -2,21 +2,11 @@
 @tailwind components;
 @tailwind utilities;
 
-#credit:hover {
-  background: #22d3ee;
-  border-radius: 12px;
-}
-.register:hover {
-  background: #a855f7;
-  border-radius: 12px;
+body::-webkit-scrollbar {
+  display: none;
 }
 
-.respawn:hover,.login:hover {
-  border: 2px solid #ffffff;
-  border-radius: 12px;
-  margin-left: -1.4px;
-}
-.home:hover {
-  text-shadow: 0 0 16px 0 rgba(255, 255, 255,0.8);
-  filter: drop-shadow(0px 0px 16px rgba(255, 255, 255, 0.8));
+body {
+  -ms-overflow-style: none;  /* IE and Edge */
+  scrollbar-width: none;  /* Firefox */
 }
